Use a different clock to obtain the start time
This code adds a new function, hdr_getnow which uses (on Linux) the
CLOCK_REALTIME clock.
The hdr_gettime uses CLOCK_MONOTONIC, which relies on an unspecificied
point in time for its initial representation. Therefore, it is not
suitable to find the actual point in time, since it may represent
elapsed time since boot or others instead of Epoch.